Add part 2 of Day 16 "Ticket translation"
This has bugs in the exhaustive enumeration, because it doesn't check
for duplicate rules in some cases.

But in the data presented, there is only one path through, and
we don't need that search logic at all.

So I'm counting this as a victory, with a proper enumeration
left for future work.  :-)